reservation-service/requirements.txt
Flask
flask-cors
psycopg2-binary
kafka-python  # Pour Kafka
SQLAlchemy  # Si tu utilises SQLAlchemy pour gérer la DB
